refactor: mongodb session management
